ProjectPointOnPlane: 1.0f and float v5 → INSN_COUNT_MATCH
Two changes:
- 1.0 → 1.0f to make MSVC emit 'fdivr DWORD' (float reciprocal)
  instead of 'fdivr QWORD' (double reciprocal). Matches ref.
- v5 typed as float instead of double so MSVC stores it as DWORD
  on the stack (when spilled), matching the ref encoding pattern.

differences are MSVC6 FPU scheduling: dot-product accumulation order
(ref does v[2,1,0], MSVC reorders to v[0,2,1] under FPU pressure)
not reachable from source.

Source-commit: 327502c3b8a39f9f92d7ea848ee2b87ee119b08e

+++ botlib/botlib.c
@@ -39520,13 +39520,13 @@ void Com_DPrintf(const char *fmt, ...) { (void)fmt; /* empty body */ }
 float *__cdecl ProjectPointOnPlane(float *a1, float *a2, float *a3)
 {
   double v4; // st4
-  double v5; // st7
+  float v5; // st7
   float *result; // eax
   float v7; // [esp+0h] [ebp-Ch]
   float v8; // [esp+4h] [ebp-8h]
   float v9; // [esp+18h] [ebp+Ch]
 
-  v4 = 1.0 / (*a3 * *a3 + a3[1] * a3[1] + a3[2] * a3[2]);
+  v4 = 1.0f / (*a3 * *a3 + a3[1] * a3[1] + a3[2] * a3[2]);
   v9 = (a2[2] * a3[2] + a2[1] * a3[1] + *a2 * *a3) * v4;
   v7 = v4 * *a3;
   v8 = v4 * a3[1];
